tab-bar: 投稿画面の表示ボタンの機能を実装しました
